Add micromamba deinit
