+++ index.js
@@ -8,20 +8,9 @@ const yaml = require('js-yaml')
 
 const cache = require('@actions/cache')
 const core = require('@actions/core')
-const exec = require('@actions/exec')
 const io = require('@actions/io')
 
-const PATHS = {
-  condarc: path.join(os.homedir(), '.condarc'),
-  bashprofile: path.join(os.homedir(), '.bash_profile'),
-  micromambaBinFolder: path.join(os.homedir(), 'micromamba-bin'),
-  micromambaExe: path.join(os.homedir(), 'micromamba-bin', 'micromamba'),
-  // Without the "-root" suffix it causes problems, why?
-  // xref https://github.com/mamba-org/mamba/issues/1751
-  micromambaRoot: path.join(os.homedir(), 'micromamba-root'),
-  micromambaPkgs: path.join(os.homedir(), 'micromamba-root', 'pkgs'),
-  micromambaEnvs: path.join(os.homedir(), 'micromamba-root', 'envs')
-}
+const { PATHS, executeSubproc, micromambaCmd, haveBash } = require('./util')
 
 // --- OS utils ---
 
@@ -32,15 +21,6 @@ function getInputAsArray (name) {
     .split('\n')
     .map(s => s.trim())
     .filter(x => x !== '')
-}
-
-async function executeSubproc (...args) {
-  core.debug(`Running shell command ${JSON.stringify(args)}`)
-  try {
-    return await exec.getExecOutput(...args)
-  } catch (error) {
-    throw Error(`Failed to execute ${JSON.stringify(args)}: ${error}`)
-  }
 }
 
 async function executeBashFlags (flags, command) {
@@ -127,10 +107,6 @@ async function retry (callback, backoffTimes = [2000, 5000, 10000]) {
   }
 }
 
-async function haveBash () {
-  return !!(await io.which('bash'))
-}
-
 function dumpFileContents (path) {
   core.info(`--- Contents of ${path} ---\n${fs.readFileSync(path)}\n--- End contents of ${path} ---`)
 }
@@ -165,10 +141,6 @@ function getCondaArch () {
     throw Error(`Platform ${process.platform}/${process.arch} not supported.`)
   }
   return arch
-}
-
-function micromambaCmd (command, logLevel, micromambaExe = 'micromamba') {
-  return `${micromambaExe} ${command}` + (logLevel ? ` --log-level ${logLevel}` : '')
 }
 
 async function executeMicromambaShellInit (shell, logLevel) {
